refactor `PedersenCommitment` and `DecryptHandle` to separate submodule
